add user interface for log tables
These tables should be filled by trigger. The log tables need to have a primary
key that is named like the log table + _id. The name of the log table must be
the name of the table to be logged + '_log'.

date_on/date_off fields in the log table are recommended. They should store the
timestamps between which the data has been valid.


git-svn-id: svn://svn.esda.local/trunk/esdaweb/vendor/plugins/scaffolding_esda@3787 28bfca7a-c10d-0410-b1fe-ac714baaa4f2